feat(blocks): riktig parallax i heron — även på iPhone
#425 lät heron falla tillbaka till statisk bakgrund på touch. Magnus ville ha
effekten på mobil också — och systerblocket bevisar att det går: transform-
baserad parallax (lager 30 % högre, translate3d en bråkdel av scrolldeltat)
fungerar överallt där bg-fixed renderar blankt i iOS Safari.

Motorn är nu en delad hook (use-parallax-background) som BÅDA blocken kör:
ParallaxSectionBlock blev av med sin inline-kopia, HeroBlock fick riktig
parallax i stället för bg-fixed. prefers-reduced-motion respekteras (lagret
står stilla). Den nu oanvända hero-parallax-utilityn ur #425 städas bort.

Grinden fångade en naken data.backgroundImage-läsning — beräkningen går nu
via heroImage-aliaset som regeln kräver.
