+++ src/components/public/blocks/ParallaxSectionBlock.tsx
@@ -1,5 +1,6 @@
 import { useRef, useState, useEffect } from 'react';
 import { cn } from '@/lib/utils';
+import { useParallaxBackground, parallaxLayerStyle } from '@/hooks/use-parallax-background';
 
 export interface ParallaxSectionBlockData {
   backgroundImage?: string;
@@ -29,8 +30,7 @@ const alignmentClasses: Record<string, string> = {
 };
 
 export function ParallaxSectionBlock({ data }: ParallaxSectionBlockProps) {
-  const sectionRef = useRef<HTMLDivElement>(null);
-  const bgRef = useRef<HTMLDivElement>(null);
+  const { sectionRef, bgRef } = useParallaxBackground(!!data.backgroundImage);
   const [isVisible, setIsVisible] = useState(false);
 
   useEffect(() => {
@@ -50,42 +50,7 @@ export function ParallaxSectionBlock({ data }: ParallaxSectionBlockProps) {
     return () => observer.disconnect();
   }, []);
 
-  // True parallax: the background layer is 30% taller than the section and
-  // translates at a fraction of the scroll delta. The old implementation used
-  // background-attachment: fixed, which is not parallax (the image just sits
-  // still) and which Chrome/iOS silently degrade to a plain static background
-  // on composited layers — the effect never fired for most visitors.
-  useEffect(() => {
-    const section = sectionRef.current;
-    const bg = bgRef.current;
-    if (!section || !bg || !data.backgroundImage) return;
-    if (window.matchMedia('(prefers-reduced-motion: reduce)').matches) return;
-
-    let raf = 0;
-    const update = () => {
-      raf = 0;
-      const rect = section.getBoundingClientRect();
-      if (rect.bottom < 0 || rect.top > window.innerHeight) return;
-      // -1 (section below viewport) … +1 (above); 0 when centered.
-      const progress =
-        (rect.top + rect.height / 2 - window.innerHeight / 2) /
-        (window.innerHeight / 2 + rect.height / 2);
-      // The layer is 30% taller, so ±15% of section height stays covered.
-      const shift = -progress * rect.height * 0.15;
-      bg.style.transform = `translate3d(0, ${shift}px, 0)`;
-    };
-    const onScroll = () => {
-      if (!raf) raf = requestAnimationFrame(update);
-    };
-    update();
-    window.addEventListener('scroll', onScroll, { passive: true });
-    window.addEventListener('resize', onScroll, { passive: true });
-    return () => {
-      if (raf) cancelAnimationFrame(raf);
-      window.removeEventListener('scroll', onScroll);
-      window.removeEventListener('resize', onScroll);
-    };
-  }, [data.backgroundImage]);
+  // Delad parallaxmotor — samma som HeroBlock (use-parallax-background).
 
   const isLight = data.textColor !== 'dark';
   const opacity = data.overlayOpacity ?? 50;
